requirements.txt
python-dotenv==1.0.1
Flask==2.3.3
requests==2.31.0
black==25.1.0
flake8==7.1.1
flask_restx==1.1.0
